guard against npm v7.3.0 brokenness, yarn prepending node brokenness

+++ scripts/relink-bin-dirs/linux.sh
@@ -16,19 +16,22 @@ mkdir -p "$RAPIDS_MODULES_PATH/.cache/cpm" \
 for DIR in $DIRS; do
     # symlink node_modules/.bin dirs to the root node_modules/.bin
     mkdir -p "$DIR/node_modules"
-    rm -rf "$DIR/node_modules/.bin"
-    ln -sf "$BIN" "$DIR/node_modules/.bin"
-    # copy the ESLint settings file (for the VSCode ESLint plugin)
-    cp ".eslintrc.js" "$DIR/.eslintrc.js"
-    # # remove and recreate the local .cache dir
-    rm -rf "$DIR/.cache"
-    # symlink to the shared .cache dir under modules
-    ln -sf "$RAPIDS_MODULES_PATH/.cache" "$DIR/.cache"
+    if [[ "$BIN" != "$DIR/node_modules/.bin" ]]; then
+        rm -rf "$DIR/node_modules/.bin"
+        ln -sf "$BIN" "$DIR/node_modules/.bin"
+        # copy the ESLint settings file (for the VSCode ESLint plugin)
+        cp ".eslintrc.js" "$DIR/.eslintrc.js"
+        # # remove and recreate the local .cache dir
+        rm -rf "$DIR/.cache"
+        # symlink to the shared .cache dir under modules
+        ln -sf "$RAPIDS_MODULES_PATH/.cache" "$DIR/.cache"
+    fi;
 done
 
-NODE_BIN_PATH="$(which node)"
+# use `which npm` because yarn prepends its own path to /tmp/yarn-XXX/node
+NPM_BIN_PATH="$(which npm)"
 NAPI_INCLUDE_DIR="$PWD/node_modules/node-addon-api"
-NODE_INCLUDE_DIR="${NODE_BIN_PATH%/bin/node}/include"
+NODE_INCLUDE_DIR="${NPM_BIN_PATH%/bin/npm}/include"
 
 # symlink node headers
 ln -sf "$NODE_INCLUDE_DIR/node/node_api.h" "$RAPIDS_CORE_PATH/include/node_api.h"
